api_yamdb/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
gunicorn==20.0.4
psycopg2-binary==2.9.3
python-dotenv==0.20.0
